project-root/backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0
openai>=1.0.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Utils
python-dotenv==1.0.1
email-validator==2.1.1
slowapi==0.1.9
redis==5.0.1

tiktoken>=0.5.0
langchain>=0.1.0
langchain-openai>=0.0.5
qdrant-client>=1.7.0
PyMuPDF>=1.23.0
python-docx>=1.1.0

# Security & Testing
safety==3.2.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-asyncio==0.24.0
